Restructured the file structure.
