+++ src/containers/dashboard/sidebar/components/UserInfo.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useContext } from 'react';
-import { UserContext } from '../context/UserContext';
+import { UserContext } from '../../../../context/UserContext';
 
 function UserInfo() {
   const {
